Issue 129, "Sticky images not freed by Show.destroy() under certain circumstances"
If there are sticky images in a feature that is part of the setup
or activated part of the current segment when a Show is destroyed,
the images weren't being unloaded (flushed) when Show.destroy() is called.

Changed show so that when destroy() is called, it de-activates
the currently active segment, if there is one.

Reviewed by Mark



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@407 e5771d8f-0841-e040-ffdb-cc069419ffcf