Issue 126, "Swapping images into fixed_image and image_sequence at runtime".
Modified ManagedImage so that you can asynchronously load an image.
This lets you tell a ManagedImage to start loading in a GRIN command,
then poll it's state each frame until it's loaded.  

Modified FixedImage to let you change the image placement, and to set a
new image.

Added code to Playground to randomly cycle through three images that get 
loaded on the fly.  This tests the new fixed_image setter methods.

Decided /not/ to allow this sort of thing for ImageSequence.  If you're
programmaticaly setting images, it's really easy to duplicate the
functionality of ImageSequence yourself, by just maintaining your
own ManagedImage[], and setting the current image of a FixedImage
from Java.


Issue 130, "Forward References for Extension Features"

It was much easier to write this if I modified Playground to take advantage
of genericgame/GameXlet, so I did.  Now there's a GrinView version of 
GameXlet, and all three GameXlet versions can offer features in common.
For now, that's just getting and setting the animation clients being animated.
This allows Playground to work completely in the genericgame framework,
and it also means that Playground is cleaner and simpler, and runs
completely under grinview.

Added a method to Lexer to get the ShowParser, and made a bunch of
the ShowParser methods public.  Also made ForwardReference public.
This makes it possible to look up forward refernces, like named features.

Tested this new functionality with Playground, by adding an ImageFrame
feature that follows a FixedImage that it refers to around on the screen.

Reviewed by Jaya



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@408 e5771d8f-0841-e040-ffdb-cc069419ffcf