Issue 131, "The fun can be halted by clones"
This issue centers around cloning assemblies...  The issue is that the
children of an assembly need not be disjoint, but the old cloning
algorigthm was assuming that the children of all features were disjoint.
OK, that last sentence is kind of dense...  Assemblies often share nodes
between their parts.  In the simplest example, consider an assembly with
two parts, where each part referred to the same FixedImage.  If you clone
the assembly, you want to end up with two nodes, and assembly and a
fixed image, in that configuration.  Under the old algorithm, however, 
the same FixedImage feature would get cloned twice, with one clone the 
first child of the cloned assembly, and the second FixedImage being 
the child of the second part of the assembly.  More generally, the 
shared descendants of the assembly could be any number of levels 
down in the tree (which is really a directed acyclic graph).  So...


First Change
------------

Split Feature.makeNewClone() into makeNewClone() and createClone().
makeNewClone()  made final, and made so that it checks if the given
feature has already been cloned; if so, it returns that clone.  If not,
it calls the protected method createClone().  createClone() needs to be 
overridden by features that support cloning.  It's much like the old 
makeNewClone(), _EXCEPT_ that Feature.makeNewClone() is now responsible 
for adding features to the HashSet

*****************************************
*    BACKWARDS COMPATIBILITY WARNING    *
*****************************************

Any GRIN clients who made cloneable extension features will get
a compile-time error, because they'll have overridden 
Feature.makeNewClone(), which is now final.  They should
change it to createClone(), make it protected, and remove code 
that adds any new child clones to the HashSet().  This latter is 
ensured by an assertion in Feature.makeNewClone().


Second change
-------------

Fixed up Feature.addSubgraph() to eliminate a potential O(n^2)
in  the presence of assemblies.  This could have manifested in
Feature.destroyClonedSubgraph().  The fix is to just check if a
feature is already in the set before adding itself and its children.


Third change
------------

Added a well-formedness assertion to Group:  Each child of a group
should have a disjoint set of children from every other child.  This
is assured for normal GRIN scene graphs by SEDoubleUseChecker in the
compiler, but this property was not checked for Group when 
resetVisibleParts() was called, e.g. to display clones.  Added an 
assertion to this effect.


Fourth change
-------------

Added a regression test to Playground for issue 131.


Reviewed by نزار and 斉藤 千尋 (Nizar and Chihiro)


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@411 e5771d8f-0841-e040-ffdb-cc069419ffcf