Issue 134, "Animation framework and falling behind in nextFrame()"
Made the animation framework do a robustness check when it's about to
skip a frame.  If the model update took at least 50% of the time allocated
to updating a frame, then instead of skipping a frame, make the model run
slower by re-setting our internal clock.  

Typically, this should have no effect, because typically any model updates will
be a small fraction of the time spent animating.  However, if the model
updates for some reason take much too long, the system will now degrade
more gracefully, with the overall animation appearing to slow down in this
exceptional case.  In the normal case on a slow player, frames will continue
to be dropped and the model will go at normal speed.

Added AnimationEngine.getModelTimeSkipped() so an AnimationClient can
discover any model time skipped in nextFrame().  This won't be used by
GRIN, but something like a game might want to use it.

Made a regression test for this new robustness feature in Playground.


Reviewed by Mark and Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@412 e5771d8f-0841-e040-ffdb-cc069419ffcf