<no issue #>  run-grinview-binary target in GenericGame
Added run-grinview-binary target to GenericGame.  This runs grinview
on a compiled binary .grin file, which is useful when developing
extension features, and their binary read/write methods.


Issue 134:  "Animation framework and falling behind in nextFrame()"

Implement a "models on probation" algorithm.

Refined the algorithm for giving up and changing the clock when model
updates are slow.  Now, a model that is slow is put on "probation,"
and is only slowed down if it's slower than a threshold value for a
second time before animation catches up.  This way, an otherwise
well-behaved model that has one bad frame will continue to run at
full speed until it catches up, at which time it will be taken
off "probation."

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@415 e5771d8f-0841-e040-ffdb-cc069419ffcf