This is a moderately complex putback.  It is completely backwards-compatible with exist projects, but it includes changes to the .grin binary file format, a significant new feature ("named commands"), and a structural change to how remote control keypresses are handled.
Issue 137, "Named GRIN commands"

Added named commands in shows, via these public interfaces:

	Show.getNamedCommand(String)	to look one up
	named_command			in show file to create one
	run_named_command		in show file to run one

This provides a facility to declare a command (or a list of commands), and
give it a name.  This can be used to declare a set of commands in the show
file, where it's natural to do so, and then use them on a show (via
Show.runCommand(Command)) from Java code, which can be in any thread.
This can also be handy when there are multiple shows, since Java code
in the director can be under the control of one show, but queue commands
to another.


Issue 141, "Remote control handling improvements (commands)"

Made Show so that it only takes out a local lock when it processes
remote control key press and release events.  These events are now
queued like commands, and executed in the animation loop.


Issue 99, "Canonicalize more stuff in the GRIN binary writer"

Now, command instances and command arrays are canonicalized.
Extension commands need to override equals(Object) and hashCode()
if they want to be canonicalized, of course.

The GRIN binary file size remains about the same, but the runtime
memory footprint should go down, particularly since simple, command
command arrays are shared.  This is particularly true if there are
many empty command arrays, or command arrays that only contain a
single command like segment_done.

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@420 e5771d8f-0841-e040-ffdb-cc069419ffcf