(no issue #):  Alternate xlet args for GameXlet
Modified GameXlet so that it checks the xlet property 
dvb.caller.parameters for xlet args, in case the GameXlet
was launched using the app launching and listing API
(see MHP 11.7.1.1).  Unfortunately, this doesn't seem
to work on all players (e.g. on the consumer PS/3 I
have at home), so a mechanism to subclass GameXlet
to look for xlet args somewhere else was added.
IXC is one possibility.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@422 e5771d8f-0841-e040-ffdb-cc069419ffcf