Issue 138: Optimized drawing and off-screen animations.
Added a special case for x, y position of the TranslatorModel being Integer.MIN_VALUE, and if so, do not add it to the animation engine's redraw region.

Added a new constant "offscreen" for the TranslatorModel's keyframe x,y position.

Added a new testcase, 8, into the Playground example to cover this case.

Reviewed by Bill.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@423 e5771d8f-0841-e040-ffdb-cc069419ffcf