<no issue #>:  Deprecation warning on linear translation not strong enough
It turns out that the deprecated "linear" style of translator doesn't
work when the child is a text element.  That's because a text element
can't calculate its upper-left hand corner without getting font metrics
from the client device, which obviously aren't available at compile time,
so the upper-left hand corner takes on the value 0.

Made the deprecation warning much more dire.  Changed the calculation 
of "linear" so that it fails on a text feature.  Made the visitor
not inherit from the abstract class, so that if new feature types are
added, we'll notice and update this bit of code accordingly.

Reviewed by Chihiro.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@425 e5771d8f-0841-e040-ffdb-cc069419ffcf