Issue 136, "Generic game should be renamed to 'GrinXlet'"
*******************************************************
*  THIS PUTBACK MAY REQUIRE A CHANGE IN YOUR PROJECT  *
*******************************************************

Changed the name of "GenericGame/GameXlet" to "GrinXlet", and moved
it to a more sensible place in the directory hierarchy.

In the GrinXlet's build files, changed name of "generic.game" property 
to "grin.xlet".  Added a failure to generic_build.xml so that if 
grin.xlet isn't set, it suggests what to fix.

Changed package name from com.hdcookbook.genericgame to 
com.hdcookbook.grinxlet.  Note that projects using GrinXlet probably
need to change this name in several places in their source code, including
in the show file.

Added to the GrinView version of GrinXlet the ability to have a
user-defined subclass.  This way, if your project defines a subclass
of GrinXlet for BD-J, you can have an se_src version of it for use
in GrinView, and safely reference your subclass from your director.

Updated SimpleGame, GrinBunny and Playground to the new GrinXlet.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@428 e5771d8f-0841-e040-ffdb-cc069419ffcf