Issue 139: Add translate/clip (and other effects) to a show's top level
Adding a mechanism to specify the top node for a rendering tree in the show file.

This commit introduces two new elements in the text show syntax.

    show_top ::= "show_top" string ;
        # Sets the name of show's top node for rendering.
        # The string should correspond to one of the feature names
        # in this show.  When the show_top element is present, then
        # a showtop_group feature should also be defined in the show.
        # This can be used to add a special effect to the show tree.

    showtop_group :: "feature" "showtop_group" name ";"
        # This can be used to give a name to the group of features
        # that represents a set of active features at runtime.
        # When a segment in this show is activated, it's active features
        # get slotted into this group.  See also "show_top".

If these elements are not present in the Show file description, then 
both showtop_group and show_top feature will point to an initially empty, 
anonymous Group.  At show runtime, when a new segment is activated, 
the ShowTop Group's part gets swapped with the segment's active features 
using Group.resetVisibleParts(Feature[]).

In the Show implementation, ShowTop is represented as a Segment with
only one element in it's activated Feature array, which is the show_top feature.
ShowTop Segment is activated during show initialization, and destroyed when
the show is destroyed.

Show's implementation as an animation client calls the showtop Segment for 
frame updates.

See Playground's case 9 for a working example.  Thanks to Bill for 
multiple reviews!







git-svn-id: https://svn.java.net/svn/hdcookbook~svn@429 e5771d8f-0841-e040-ffdb-cc069419ffcf