Issue 132, "Add include directive"
Added an include directive to show file (and mosaic file).  It's
implemented in the lexer, and allows "$include filname.txt ;" anywhere
inside a file.

Issue 133, "Enhanced tree processing in GRIN compiler"

Added a method SENode.postProcess(ShowBuilder).  This is called
on every show node after the show tree is parsed, but before the
double use checker is run.  Withing postProcess(), a node can
add child nodes, or add a parent to itself.

Added SENode.changeFeatureReference(Feature from, Feature to).
All SENode types must implement this.  It's needed for the feature
of adding a parent to the node, because most references to the child
need to be changed to refer to the parent when this happnes.

Added a section about extending GRIN to the GRIN design document.
This includes a description about the compiler, the extension parser,
and having a SE version of an extension.

Added a "bouncing arc" feature to Playground, in order to test tree
processing.  Bouncing arc adds a Translator as its parent, and adds
other nodes needed to create a bouncing effect.

Added visitUserDefinedModifier to SEShowVisitor, for completeness.
This might make it slightly easier to add a new modifier type,
since it makes the need to visit the modifier's part more explicit.

Changed text name of show top segment in GrinView to " $show_top$ ".



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@430 e5771d8f-0841-e040-ffdb-cc069419ffcf