Fixed a bug running GrinView on Windows.
It appears that GrinView couldn't be run from within ant on Windows.
It worked fine on Mac, but on Windows it would go into deadlock as
soon as you tried to do keyboard input (with the focus on the Xlet
window).  It turns out that the ant/windows combination was unhappy
with GrinView trying to read from stdin.  Reading commands from stdin
was a legacy thing, from before GrinView had a GUI to drive it, so
I just made GrinView default to ignoring stdin.  Now, GrinView works
great on Windows, so "ant grinview" in any GrinXlet project is perfectly
fine on Windows now, as it has always been on Mac.

Also, fixed an obscure bug with image scaling that only would show
up when using the GrinXlet grinview-jar file.




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@435 e5771d8f-0841-e040-ffdb-cc069419ffcf