<no issue #>
Made an ImageManager method to get a ManagedImage directly from a URL.
This is handy for using the reference counting of ManagedImage with
images that come from places like web servers.

Added getFont() to com.hdcookbook.grin.features.Text.  This makes
it possible for xlets to do the font metrics calculation needed to
break text up.  Noted that for font strips, it returns null.

Added toString() methods to most SE<...> GRIN features, commands and
whatnot.  This makes the browser display in GrinView easier to read.

Added some demos.  The demo directory is a collection of unmaintained,
stand-alone projects that aren't integrated into the build.

Reviewed by Jaya



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@441 e5771d8f-0841-e040-ffdb-cc069419ffcf