No issue # : Making a BumfGenerator a part of the Disc Creation Tools suite.
BumfGenerator is a simple command line tool that takes an existing BDMV directory and a few other things as parameters, and then generates the bumf xml file
which can be used for performing VFS update with that BDMV directory as new files.

One can use this tool to generate the bumf file, then sign the xml using hdcookbook's security tool and obtain the .sf file corresponding to the xml.  Note that security tool also allows the generation of the bu.discroot.cert.

The BumfGenerator tool also takes an optional parameter for progressive playlist items, as comma-separated 5 digit numbers.  If this parameter is specified, then the corresponding m2ts files in the BDMV directory will be listed as a progressive playlist assets in the bumf xml file instead of the standard assets. 

Adjusting the BridgeheadXlet build to work with the new tool location.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@454 e5771d8f-0841-e040-ffdb-cc069419ffcf