Issue 146: Progressive playlist support (partial - test case only).
Adding a simple example to the hdcookbook repository that uses a progressive playlist feature.  This example simply unzips the files needed for preforming VFS update to the BUDA area, does the VFS update, enables the clip that's listed as a progressive playlist asset in the BUMF, and finally, plays back the clip.

BUMF.xml and BUMF.sf used for the VFS update are generated using hdcookbook's bumfgenerator tool and the security tool during the test project's build time.

Thanks to Bill for testing this out on PS3 and a review.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@455 e5771d8f-0841-e040-ffdb-cc069419ffcf