Fix for issue 143, "Image Cloning Bug"
There were two bugs here.  The first was that cloned features need to
do the equivalent of setup/unsetup.  This only comes up with images
(of course), and it doesn't show up if the feature you're cloned
from stays active.  The fix was to do a prepare() of the ManagedImage
in the cloned feature when it's created, and a balanced unprepare()
when it's destroyed.

The second bug was an interaction betwen FixedImage.replaceImage()
and cloning a fixed image feature.  The clone method was using the name
of the original image, rather than the image that replaced it; this was
fixed.

Added a regression test for the first of these bugs to Playground.

Added documentation to Feature methods related to cloning to explain
the needed lifecycle for the first bug.

<No issue #>

Added a simple example script showing show_top.  Fixed up some
small build issues.

Reviewed by Jaya.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@459 e5771d8f-0841-e040-ffdb-cc069419ffcf