Fix for the issue 145. Profiling of Xlets
In this fix a bunch of profiling methods are added to
the Debug class for time profiling the xlets. Xlets can make
calls to these methods to time profile their methods. 
The puback includes a small program that runs on a
remote computer and listens to the profiling commands sent
over an UDP socket by an xlet running on a Blu-ray player.

Reviewed by Bill


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@460 e5771d8f-0841-e040-ffdb-cc069419ffcf