Issue 152:  Key typed events.
Extended RCKeyEvent, Show, Segment and related classes so that GRIN
can be extended to support key typed events.  This allows the GRIN event
dispatching mechanism to be used, if your xlet has a source of key
typed events (either from an external keyboard, a virtual keyboard, or
some other support).  GRIN does not include built-in support for key
typed events, but with these tweaks it can be extended to support them.

<no issue #>

Made the clean target of profiler a little more assertive, but deleting
the entire build directory.

Reviewed by Mark



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@461 e5771d8f-0841-e040-ffdb-cc069419ffcf