This is a GRIN based yahoo weather widget for the Blu-ray platfom. It is based on the Java FX weather widget available at:
	http://www.javafx.com/samples/WeatherWidget/index.html

The forecast information is obtained by pressing the down arrow
button.

The left arrow key can be used to erase the zip code displayed in
the left upper corner. After entering the desired code, pressing the
enter key will fetch the weather data for the specified city.

This will be made more clear in the next update by visual means.

Reviewed by Bill



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@468 e5771d8f-0841-e040-ffdb-cc069419ffcf