Issue 154:  Make SFAAEngine for GRIN animation framework.
Second putback for this issue:  Added some debugging output
to PlayerWrangler, to make it easier to debug video playback
issues.  Fixed a small bug in SFAAEngine.  Fixed debug stack
trace in AnimationEngine, and made destroy() logic easier
to follow.

SFAATest:  Added ability to set number of buffers used by the
SFAA (between 1 and 4 may be set with the remote control).



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@472 e5771d8f-0841-e040-ffdb-cc069419ffcf