Partial addressing of issue 158:  Font strip optimizations
This is a pretty simple putback that spends 4 bytes/character in
order to avoid allocating a bunch of heap objects each time font
strip text is displayed.  Originally we weren't to worried about
this, since we expected font strip text to be stationery most
of the time, but there have now been reports of wanting to animate
font strip text, too.

More could be done here:  The perfect hash speedup mentioned in
the issue is alluring, and for strings known at compile time, there's
no real reason to not generate the CharMapInfo[][] at compile time,
and just record it as a bunch of integer indexes into an array
maintained while reading the binary file.

Reviewed by Jaya



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@478 e5771d8f-0841-e040-ffdb-cc069419ffcf