Issue 161 - "Flash" when a segment's setup clause does real loading
Made it so that the show top group's visible features are set to
the features in the active clause of the segment as soon as those
features are all set up.  Before, it was waiting until the features
in the setup clause were also being loaded.

This just required re-arranging some code.  On a correctly written
show, where a segement's features are all set up before the segment
is entered, this will have a trivially small performance impact.

Reviewed by Mark.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@484 e5771d8f-0841-e040-ffdb-cc069419ffcf