<Partially addresses issue 150>:  "Browser for profiling data"
This putback adds a Crude, prototype  profiling data capture & 
display program.  It displays in a GUI, but it's command-line
driven. We expect that this will be substantially revised in 
the future; this version is just good enough to do some 
exploration and make some slides for a JavaOne talk.

Also Added a thread ID argument for profiling.  Put profiler calls in
a few key places in the animation manager and in the setup manager.

Reviewed by Jaya



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@494 e5771d8f-0841-e040-ffdb-cc069419ffcf