Issue 171, "BridgeheadDiscImage should be able to download larger files"
Also, other misc. fixes and enhancements targeted at the disc image
for the HD cookbook booth at JavaOne.

Changes to bridgehead xlet:  Now, instead of a .zip file, it 
streams a ".hdcvfs" archive.  This is an uncompressed format where
each file is sent as a filename, a byte count, and then the contents
of the file.  This way, the bridgehead xlet doesn't need to have a
copy of the entire archive in memory, so now bridgehead can be used
to download images that contain large files, like video assets in the
tens or hundreds of MB.

Also changed the name of the BUDA manifest file from "sample.xml" to
"manifest.xml".

Added an "images_to_skip" clause to mosaic descirption files, so you
can set certain images to not be included in the image mosaic.

Made XletLogger so it follows the last line, when the last line is on screen.

Changed bumfgenerator so that it copies the BDMV directory being uploaded
to a flat directory that contains only 8.3 file names.

Fixed up the bridgehead xlet and the progressive playlist demo to use the
new bufmgenerator.  Added meta-data to the bridgehead disc image (including
a cute image) for presentation on the players' BUDA maintenance screen.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@502 e5771d8f-0841-e040-ffdb-cc069419ffcf