<no issue #>:  Fixes uncovered in testing with a PC and a certain player
The debug log output was unreadable when using the Windows command-line
version of telnet, because it treated \n as newline without carriage
return.  Added \r to the telnet output.

The bridgehead's PRF was unreadable on a certain player, perhaps
because the last line lacked EOLN?

On a certain player, we got a socket reset exception on uploading to
the bridgehead if the PC's socket was closed too soon.  Now they
do a handshake with the socket going the other way.

Not reviewed, and sorry, no webrev - I don't have svnwebrev installed
on my PC :-(



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@503 e5771d8f-0841-e040-ffdb-cc069419ffcf