A sample program with socket connection timeout functionality.
With Personal Basis Profile 1.0, there is no way of setting the Socket timeout
when attempting connection to a host. Some players have been observed to have long socket timeouts.
With long socket timeouts it's often hard to distinguish between non-availability of the remote host,
a genuine network problem or network configuration error either at the remote host or with the
player/application itself.

The Socket timeout functionality is simulated by creating a thread that manages the timer,
and signals appropriately to the socket connection thread upon timeout.

Reviewed by Bill


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@513 e5771d8f-0841-e040-ffdb-cc069419ffcf