Issue 24, "Go Lakers!"
How about that Fish, with two (count 'em) clutch 3-pointers
at the end of the game?  That's what championships are made
of.  Orlando's not going to go quietly, though, so I'm
sticking with Barack Obama's prediction:  Lakers in 6.


Issue 144, "DebugLog can conflict with self on port 6000"

Made server socket keep trying over a period of two seconds.


Issue 108, "GrinView deadlock when watching drawing"

Made GrinView so it posts a command to the show under test
to handle mouse events.  This cures the deadlock, because it's
always safe to post a command.  The same is already done for
remote control events.  In a real xlet, mouse events don't
need to be queued like this - it's OK to just stall the mouse
thread (if it exists) until the show is in a state where the
model can be updated.


Issue 115, "ImageSequence could use get/set currFrame"

Added a getter and setter method.  This makes an image sequence an
easier way of programmatically selecting an image, without needing to
use ManagedImage and a fixed_image.

Added a method to Clipped to change the clip region.

Tested both in Playground.


Issue 153, "More Easing - list of points..."

Added syntax to a show to let you specify a list of points.
Used this in Playground for a new test that shows the moon
orbiting the earth in a circle.  Results were good - 360 data
points (describing the circle) get compressed to 37 linear
interpolations when max-error is set to 2.

This feature might be useful for a visual tool that lets the
path of an object be described by some kind of formula, like
a spline curve.  As long as you can generate a list of points
at compile time, you can efficiently store the path, and
animate it at runtime.


<no issue #>:  Bug in popSegment

Fixed an assert statement that was overly aggressive.


<no issue #>:  Translator and Integer.MIN_VALUE

Workaround of a bug in desktop Java on the Mac, if you can
beleive that!  When a translator is set to offscreen,
it was causing a translation of Integer.MIN_VALUE to
be applied to a Graphics object.  Bizarrely, that caused
a later SrcOver draw to the same buffer to fail.

Reviewed by Jaya


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@514 e5771d8f-0841-e040-ffdb-cc069419ffcf