Issue 168, "Tracking of image load/unload is desirable"
BufferedImage already had an AssetFinder method for when the
image is no longer needed:  AssetFinder.destroyImageBuffer().
It was already being called from DirectDrawEngine, but not
from RepaintDrawEngine, so this putback fixes it to be called
from RepaintDrawEngine too.

Added AssetFinder.notifyLoaded(ManagedFullImage) and
AssetFinder.notifyUnloaded(ManagedFullImage).  They are
called from ManagedFullImage.


Issue 148, "GrinView still image to fake video background by segment"

Added an extension to the Show file syntax that lets you specify
a list of background images.  In grinview, that list shows up in the
GUI just above the list of segments.  Double-clicking a background
image name sets the new background, and stops animation (setFps(0))
to make the background visible.

Reviewed by Jaya



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@518 e5771d8f-0841-e040-ffdb-cc069419ffcf