<no issue # - fixing recent regressions>
Fixed a setup bug in text feature where setup wasn't happening properly
if script called Text.setText before the text feature was setup.  This 
could have caused setup_done to not be triggered.

Restored Text.getFont() method that was taken away in recent putback.
It is needed by xlets, e.g. to calculate font metrics.  Indeed, this
is done in twitterGRIN.

Issue 149:  Scaling shows at compile time ("QHD debug screen would be nice")

Implemented compile-time scaling of GRIN shows, so a full HD show can be
automatically scaled down to 720p HD, or QHD, or even SD.  Tested by
making a new version of GrinBunny that runs at different scales,
in xlets/tests/functional/ScaleBunny.

In the grin compiler, created new SEScalableNode interface, to be implemented
by all RC handlers and features that support scaling.  Implemented scaling in
GRIN core nodes, e.g. by scaling down the placement of images.  For Text,
just scaled down the font size - there's no way to fix the aspect ratio
when xScale != yScale.  In the GRIN compiler, scaling is done as a pass
toward the end.  If any of the show's nodes aren't scalable, and a scale
factor (other than 1.0) is being applied, then compilation fails.

Made box so that at runtime it supports different widths for the horizontal
and vertical lines.  This is needed for scaling that changes the pixel
aspect ratio.

Implemented scaling in the visual RC handler by changing the rectangle
used for mouse support.

In the mosaic compiler, made it look at the images, and the placements of
all of the features that use the images.  If the placement is smaller than
the source image, reduce the size of the image what's needed by the features,
thus making for smaller matrixes when scaling is done.

Added Show.scale(), Show.getXScale(), Show.getYScale(), Show.getXOffset()
and Show.getYOffset() for use by directors.  Generally, directors should 
never get coordinates from a scaled show - they should maintain them as 
instance variables, sized to the "native" screen size of the xlet (probably 
full HD).  If they do this, and then scale those coordinates every time 
they set a coordinate in the show, all will be well when things are scaled.

Changed the font strip support so that the character metrics are closer to
what's needed for rendering, rather than being based on coordinates in the
design image.  This is equivalently efficient at runtime, but in the new
form it's easier to know what to scale by which scale factor in the compiler.
Implemented scaling of fontstrips in the font strip compiler.

Implemented GRIN compiler scaling in SEFontStrip.  Note that it expects, but
does not enforce the same scaling factor between the two compilers.

Added ScaleBunny show to scale the grin bunny game to 720 HD, QHD, and SD.

Implemented debug.screen.scale property in GrinXlet, so that the debug screen
can be scaled down.  Used this new property in weatherWidget and twitterXlet,
so now the debug screen is sized for the screen configuration.

Reviewed by Jaya



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@520 e5771d8f-0841-e040-ffdb-cc069419ffcf