This commit is toward a branch; please ignore.
Adding two new methods to GrinXlet, pushKeyInterest(Show) and popKeyInterest().
They can be used to tell GrinXlet the stacking order in which KeyEvents 
should be sent off to.  GrinXlet holds an array of Shows, and send KeyEvents
to the first Show in the array, and if that show doesn't consume the key, then
to the second Show, and then to the third Show, so forth.
If the Show's active Segment have an RCHandler that uses the Key, then the
Key is "consumed" by that Show.

MouseEvents are currently sent to all the shows in the keyInterest table.

Also updating GrinViewer's GrinXlet to handle KeyEvents and MouseEvents, so that
multiple shows will get key and mouse events under the GrinView run.

Finally, removing Box.setResized() method after talking with Bill.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@545 e5771d8f-0841-e040-ffdb-cc069419ffcf