Fix for 183, "Grinview messes up alpha blending on Mac Intel with Java 5"
It turns out that this is a bug in the Apple quartz library.  Changed
GrinView to deteck Mac/Intel/Java 5.  When it does, it issues a
message to stdout, and uses a system property to turn off Quartz and
revert to the Sun renderer.  Thanks, Jon, for tracking down the
Apple issues and the system property!

Reviewed by Jon 



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@549 e5771d8f-0841-e040-ffdb-cc069419ffcf