Issue 188:  Allow tuning of collapsing draw targets
The draw target collapsing algorithm would combine two draw targets
if doing so cost less than 40,000 pixels.  That number was a guess.
This putback lets an application tune that value.

One reason this is valuable is to set it to -1, for situations where
an xlet wants to disable the optimization completely in order to get
consistent and predictable animation performance.

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@564 e5771d8f-0841-e040-ffdb-cc069419ffcf