Issue 191:  Enhance "watch drawing" in GrinView; change box programmatically
Changed the "watch drawing" mode in GrinView when a framerate is set.  Now,
it doesn't force you to single-step through the drawing.  Instead, it behaves
like the debug GrinXlet, and just holds on the erase view for 1/4 frame,
on the draw view for 1/4 frame, then proceeds.  The old behavior of 
single-stepping through is naturally available by setting the framerate
to zero.

The UI state model is a little weird and should be re-designed, but that
would be a more substantial change.  For now, it just forces watch drawing
off when the "stop" button is pressed to avoid a potential deadlock.


Enhanced the box feature so other parameters can be changed.  From measurement,
adding methods to an existing class doesn't increase load time noticably, so
the marginal value of this change is still justified.  Also, any decent
obfuscator will optimize away these new methods anyway in a deployed xlet.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@566 e5771d8f-0841-e040-ffdb-cc069419ffcf