Issue 193:  BUDA credentials not working
They work now, though a small number of players seem to have problems
with credentials containing wildcards.

This putback removes the newlines that were in the base-64 encoding of
the XML tags holding the credential in the PRF.  That was what was causing it to
fail.  Note that this may be a player bug; base-64 encoding allows newlines.

Also, this putback adds some whitespace after each new tag in the PRF,
so it looks nicer.

This putback also contains some miscellaneous unrelated enhancements to the
DataAreaAccessTest xlet.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@571 e5771d8f-0841-e040-ffdb-cc069419ffcf