Issue 199:  ManagedFullImage has unfriendly failure mode if unget precedes unprepare
Added a warning to ManagedFullImage in this case, triggered when
Debug.LEVEL > 0 and the image has been loaded.

Changed ManagedFullImage to store image width and height as instance
variables, so that image.getWidth() and image.getHeight() don't need to
be called.  That way, if width and height are consulted after image.flush() 
(as happens when unget precedes unprepare), there is no danger of
triggering a re-load of the image.

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@580 e5771d8f-0841-e040-ffdb-cc069419ffcf