206: Playlist tool IN_time/OUT_time Integer overflow
Changing the following fields from int to long

- PlayItem's inTime and outTime 
- SubPlayItem's inTime, outTime and syncStartPtsOfPlayItem,
- Mark's markTimeStamp, duration

and adding UnsignedIntHelper class to help the i/o of converting 4 bytes to long and back.

Please note that if you have the playlist in the xml format converted with earlier playlist prior to this bug fix, and have one of the above fields in negative number, then the updated tool will translate those values differently from what the original .mpls had.  

Many thanks to Gunnar Adler for reporting the issue, testcase and reviews. 




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@587 e5771d8f-0841-e040-ffdb-cc069419ffcf