<no issue #>:  Off with their heads!  Headless support for GRIN compiler
Made the grin compiler so that it works in headless mode, and made that
the default.  Now, by default, the grin compiler won't create a Frame that
it uses to show mosaic construction.  To get that old behavior back, add
the command-line argument "-show_mosaic".

Added "-show_mosaic" to the bookmenu project, so this will get tested in
the regression test ("ant all").

Reviewed by Chihiro.  Thanks!




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@589 e5771d8f-0841-e040-ffdb-cc069419ffcf