Issue 210:  Bug cloning Text features
When a setup text feature was clone, the width was copied over, but
the setupMode flag was not.  This made it so that subsequent setText
operations failed to calculate the font metrics, which caused the feature's
damage rectangle to not be re-calculated.  Fixed.


Issue 210:  Profiling improvements

Added a devastatingly cool scale display to the simple PCProfiler GUI.
Put updated images in www/profiler.html.

Added constants PROFILE_ANIMATION and PROFILE_SETUP to Debug, and defaulted
them to false.  These now control profiling of the animation loop and the
setup loop.  This is handy if you want to turn off this profiling, but
you want profiling on to measure something else, because now you won't get
these extra UDP packets adding noise to the measurement you care about.
Also, when displaying you don't have to filter out all the animation loop
stuff.

Slight, unrelated  debugging improvement to GrinXHelper:  better toString().




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@601 e5771d8f-0841-e040-ffdb-cc069419ffcf